+++ inst/doc/vignette.R
@@ -1,9 +1,0 @@
-## ---- include = FALSE---------------------------------------------------------
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-knitr::opts_chunk$set(warning = FALSE, message=FALSE, error=TRUE, purl=FALSE, cache=TRUE)
-
-knitr::opts_chunk$set(dpi=100)
-