+++ NEWS.md
@@ -1,3 +1,12 @@
 # covsim 1.0.0
 
 Third release. Accompanies journal publication in JSS. 
+
+# covsim 1.1.0
+
+rIG: The A matrix is now triangular by default
+
+rPLSIM: Added tests for whether an identical pair has been calibrated. 
+This speeds up rPLSIM if sigma.target has many identical values, and skewness and kurtosis values are identical. 
+
+vita: Now checks margins are correctly named.